% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BCa_bootstrap_ci}
\alias{BCa_bootstrap_ci}
\title{Estimate bias and skewness corrected bootstrap confidence intervals}
\arguments{
\item{bootstrapped_parameter_estimates}{A \code{numeric vector} storing bootstrapped estimates of the target parameter. A couple of missing values are allowed and will be removed if present.}

\item{jackknife_parameter_estimates}{A \code{numeric vector} containing Jackknife estimates of the target parameter. A couple of missing values are allowed and will be removed if present.}

\item{original_parameter_estimate}{A \code{double} representing the original point estimate of the target parameter. A missing value is not allowed.}

\item{level}{A \code{double} between 0 and 1, indicating the desired confidence level for the estimated BCa bootstrap confidence interval of the target parameter. Defaulting to \code{0.95}.}

\item{silence}{An \code{integer} controlling the level of progress reporting for debugging and analysis. Set to \code{-1} or \code{0} for detailed reports. Defaults to \code{1} to suppress output.}
}
\value{
A \code{numeric vector} with two entries: the lower and upper bounds of the BCa bootstrap confidence interval, respectively.
}
\description{
This function estimates the bootstrap confidence intervals for a specified parameter using a Bias-Correction and Acceleration (BCa) approach. It requires precomputed Jackknife (leave-one-out) estimates of the parameter of interest. These can be generated by iterating over all unique SampleIDs in your dataset and employing the \code{leave_one_out()} function from this package for each unique ID.
}
\details{
Estimate bias and skewness corrected bootstrap confidence intervals

To accurately estimate the theoretical confidence interval for the parameter of interest, the original dataset must adequately represent the underlying distribution of the population parameter. Typically, larger datasets increase the likelihood of representativeness. To compute bootstrap confidence intervals independent of Jackknife estimates, consider using the \code{bootstrap_ci()} function instead.
}
\examples{
\dontrun{
  fictional_original_estimate <- 2.1
  fictional_bootstrap_estimates <- runif(n = 1e3, min = 0.5, max = 2.5)
  fictional_jackknife_estimates <- runif(n = 1e1, min = 1.5, max = 2.5)
  BCa_bootstrap_ci(fictional_bootstrap_estimates,
                   fictional_jackknife_estimates,
                   fictional_original_estimate,
                   level = 0.95,
                   silence = 1)
}
}
