% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CharacterEstimatePrecision}
\alias{CharacterEstimatePrecision}
\title{Estimate MS precision measures}
\arguments{
\item{data}{A data frame or table with format LFDT converted to list so that each column is a list element}

\item{silence}{- Integer, hide helpful (maybe unhelpful for most people) progress reports. Set to 0 for these progress report to appear. For additional debugging information, set this to \code{-1}.}
}
\value{
A list containing estimated MS precision estimates for one group. Can be melted together by using \code{setDT()} from the \code{data.table} package
}
\description{
Estimate precision of EQA data when the ID columns are of string type. All ID columns must be character vectors for this function to work
}
\details{
Estimate MS precision measures (variances, CVs and lambda) for character IDs

These estimates are used as part of calculation of relative selectivity, and construction of prediction intervals. They can also be used to visualize the the MS precisions using the much used CVs in percent
}
\examples{
\dontrun{
  print(1)
}

}
