% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_absolute_dins}
\alias{estimate_absolute_dins}
\title{Calculate Absolute Differences in Non-Selectivity}
\arguments{
\item{data}{\code{list} or \code{data.table} - The input dataset that must contain the following columns: \code{SampleID}, \code{ReplicateID}, \code{MP_A}, and \code{MP_B}.}

\item{silence}{\code{integer} - The verbosity level of the output. Be aware that excessive progress reporting might negatively impact performance. The valid input options are as follows:
\itemize{
\item{\code{1: }}{All progress reports are silenced (default setting)}
\item{\code{0: }}{Progress reports are partially shown, while debugging reports are omitted}
\item{\code{-1: }}{All progress reports are enabled, providing the most detailed output}
}}
}
\value{
A list containing the point estimate of the absolute magnitude of differences in non-selectivity with a precision of 1e-6.
}
\description{
This function calculates the absolute value of differences in non-selectivity, which are crucial when evaluating the commutability of reference materials or external quality assessment materials in the context of in vitro diagnostic medical devices.
}
\details{
Calculate Absolute Differences in Non-Selectivity

Differences in non-selectivity between diagnostic devices may lead to difficulties in commutability assessments. Larger values of absolute differences are particularly problematic. The function provides a point estimate of these differences with a precision of 1e-6.
}
\examples{
\dontrun{
library(fasteqa)
data <- simulate_data_eqa(list(n = 25, R = 3, cvx = 0.06, cvy = 0.04))
calculate_absolute_dins(data)
}
}
