% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_zeta}
\alias{estimate_zeta}
\title{Estimate differences in non-selectivity with zeta}
\arguments{
\item{data}{\code{list} or \code{data table} - Data with elements/columns \code{SampleID}, \code{ReplicateID}, \code{MP_A} and \code{MP_B}}

\item{silence}{\code{integer} - Controls verbose. Allowing verbose, will slow down estimation of zeta. Thus, in general, it should not be used. There are three valid inputs:
\itemize{
  \item{\code{1: }}{All progress reports are silenced. This is the default.}
  \item{\code{0: }}{Estimation steps and temporary results are printed to the console}
  \item{\code{<0: }}{Debugging. Expert use only.}
}}
}
\value{
A \code{list} containing the a point estimate of differences in non-selectivity, zeta. The estimated zeta value is a floating-point value, meaning that the precision is 1e-6 (six decimals precision).
}
\description{
Estimate the magnitude of differences in non-selectivity, zeta. The measure, zeta, is a theoretical measure being the ratio of the pooled prediction error variance of the OLS model and the irreducible variance of the OLS model.
             This function estimates zeta using a plug-in estimator for it. See references (coming..) for details on the estimator.
}
\details{
Estimate differences in non-selectivity with zeta

Differences in non-selectivity between in vitro diagnostic medical devices may cause problems in e.g., evaluation of commutability of external quality assessment materials or certified reference materials. A large estimate of zeta indicates that we have considerable differences in non-selectivity between compared IVD-MDs.
         An upper limit of acceptable zeta estimates may for example be determined based on an particular allowable average relative increase in prediction interval widths due to differences in non-selectivity. Recommendations for choosing allowable average relative increase in prediction interval widths may be found elsewhere.
}
\examples{
\dontrun{
  library(fasteqa)
  data <- simulate_data_eqa(list(n = 25, R = 3, cvx = 0.06, cvy = 0.04))
  estimate_zeta(data)
}

}
