% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Estimatek}
\alias{Estimatek}
\title{Estimate MS precision measures}
\arguments{
\item{data}{A data table or a data frame converted to a list so that each column is a unique list element. The ID columns must be of character type and contain SampleID, ReplicateID, MP_A and MP_B}

\item{silence}{An integer (-1, 0, 1) that controls how much is printed. 0 shows some messages regarding temporary output, whereas -1 shows even more messages for debugging. -1 is used for superusers.}
}
\value{
A single estimate of k (in a list) based on the specific grouped data. The k values are float values meaning that the precision is 1e-6 (six decimals precision). This is considered enough for us
}
\description{
Estimate the relative non-selectivity measure k, that is the ratio of the pooled prediction error variance and the sum of repeatability variances.
}
\details{
Estimate relative non-selectivity measure k based on data

Differences in non-selectivity between measurement systems may cause problems in e.g., evaluation of commutability. A large value of k indicate that we have have large differences in selectivity
}
\examples{
\dontrun{
  print("Estimate k")
}

}
