% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fun_of_replicates}
\alias{fun_of_replicates}
\title{Estimate a Statistic for Each Sample}
\usage{
fun_of_replicates(data, fun = "mean")
}
\arguments{
\item{data}{A \code{data.table} or \code{list} object. Must contain
\code{SampleID}, \code{ReplicateID}, \code{MP_A} and \code{MP_B}.}

\item{fun}{A \code{character} string. Which statistic is to be calculated for
each SampleID. Possible choices include \code{mean}, \code{var} (variance),
\code{sd} (standard deviaton), \code{cv} (coefficient of variation),
\code{median}, \code{min} (minimum) and \code{max} (maximum).}
}
\value{
A \code{list} with elements \code{SampleID}, \code{MP_A} and \code{MP_B}.
\code{MP_A} and \code{MP_B} contains the estimated sample-wise statistics.
}
\description{
For each sample, uses the replicated measurements to estimate
a given statistic.
}
\details{
This function handles \code{NA}-values automatically. If all replicated measurements
for a given sample are \code{NA}-values, the corresponding estimated statistic will also
be \code{NA}.
}
\examples{
\dontrun{
  library(fasteqa)
  test_data <- simulate_eqa_data(list(n = 25, R = 5, cvx = 0.01,
                                      cvy = 0.01, cil = 2, ciu = 10))
  print(as.data.frame(fun_of_replicates(test_data, "mean")))
  print(as.data.frame(fun_of_replicates(test_data, "var")))
  print(as.data.frame(fun_of_replicates(test_data, "cv")))
}
}
