% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glucose_cs_data}
\alias{glucose_cs_data}
\title{Glucose Serum Clinical Sample Meausurements}
\format{
## `test_data`
A data frame with 75 rows and 6 columns:
\describe{
  \item{SampleID}{ID of clinical sample}
  \item{ReplicateID}{ID of replicated measurements}
  \item{Advia}{Measurement results from an IVD-MD named Advia.}
  \item{Alinity}{Measurement results from an IVD-MD named Alinity.}
  \item{Cobas}{Measurement results from an IVD-MD named Cobas.}
  \item{Vitros}{Measurement results from an IVD-MD named Vitros.}
}
}
\source{
Noklus
}
\usage{
glucose_cs_data
}
\description{
Glucose Serum Clinical Sample Meausurements
}
\keyword{datasets}
