% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glucose_eqam_data}
\alias{glucose_eqam_data}
\title{Glucose Serum External Quality Assessment Material Sample Meausurements}
\format{
## `test_data`
A data frame with 9 rows and 6 columns:
\describe{
  \item{SampleID}{ID of EQA material sample}
  \item{ReplicateID}{ID of replicated measurements}
  \item{Advia}{Measurement results from an IVD-MD named Advia.}
  \item{Alinity}{Measurement results from an IVD-MD named Alinity.}
  \item{Cobas}{Measurement results from an IVD-MD named Cobas.}
  \item{Vitros}{Measurement results from an IVD-MD named Vitros.}
}
}
\source{
Noklus
}
\usage{
glucose_eqam_data
}
\description{
Glucose Serum External Quality Assessment Material Sample Meausurements
}
\keyword{datasets}
