% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kurtosis}
\alias{kurtosis}
\title{Calculate Sample Excess Kurtosis of a Random Sample}
\arguments{
\item{x}{A \code{numeric} vector. A random sample.}

\item{na_rm}{A \code{logical} value. If \code{TRUE}, \code{NA}-values are
removed prior to calculation of the sample excess kurtosis.}
}
\value{
A \code{double}. The calculated sample excess kurtosis.
}
\description{
Calculates the sample excess kurtosis of a random sample \code{x}, using
the adjusted Fisher-Pearson standardized moment coefficient estimator.
}
\details{
Calculates the sample excess kurtosis of a random sample
\eqn{\lbrace x_i \rbrace_{i=1}^{N}}. The sample excess kurtosis, attempts
to estimate the theoretical excess kurtosis, \eqn{\Kappa}, by using
the following estimator (adjusted Fisher-Pearson standardized moment
coefficient):

\eqn{\hat{\Kappa} = \frac{N(N+1)}{(N-1)(N-2)(N-3)} \cdot \frac{\sum_{i=1}^{N}(x_i - \overline{x})^4}{s^4} - 3 \cdot \frac{(N-1)^2}{(N-2)(N-3)}}.

Here, \eqn{s^2} is the unbiased sample variance.


Note: this estimator will be biased if not
\eqn{x_i \sim \mathrm{N}(\mu, \sigma^2)}. Moreover, \eqn{\hat{\Kappa}}
estimates the theoretical excess kurtosis and not the raw kurtosis. Add
\eqn{3} to the sample excess kurtosis to get the sample raw kurtosis.
}
\examples{
# Reproducibility
set.seed(99)

# Simulate from one million observations from N(0, 1)
y <- rnorm(n = 1e6)
sample_excess_kurt <- kurtosis(y)

# Compare with theoretical excess kurtosis of N(0, 1)
cat("Theoretical excess kurtosis:", round(0, 3), "\n",
    " Empirical excess kurtosis:", round(sample_excess_kurt, 3), "\n")

}
