% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{leave_one_out}
\alias{leave_one_out}
\title{Leave-one-out on clustered EQA clinical sample data}
\arguments{
\item{data}{A \code{list} or \code{data.table}. Must contain
\code{SampleID}, \code{ReplicateID}, \code{MP_A} and \code{MP_B}.}

\item{loo_id}{An \code{integer}. The \code{SampleID} that is left out.
Default value is \code{1}.}
}
\value{
A \code{list} containing the original data, but without the sample id corresponding to the given \code{loo_id}
}
\description{
Needed to calculate jack knife estimates of a parameter, that is required when using \code{BCa_bootstrap_ci()}. Alternatively one could use the Jack knife estimates to calculate standard error or bias of the estimator of relevance
}
\details{
loo_ids can not be vectorized directly in R. Use \code{sapply()} or \code{replicate()} to leave out sample IDs one by one
}
\examples{
# Get data without the fifth sample
print(as.data.frame(leave_one_out(test_data, 5L)))

}
