% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loo_global_precision_estimates}
\alias{loo_global_precision_estimates}
\title{Leave-One-Out (LOO) Imprecision Estimates Based on Clustered EQA Clinical Sample Data}
\arguments{
\item{data}{A \code{list} or \code{data.table}. Must contain
\code{SampleID}, \code{ReplicateID}, \code{MP_A} and \code{MP_B}.}

\item{loo_id}{A \code{integer}. The ID of the \code{SampleID} to be
excluded before calculating the imprecision estimates.}
}
\value{
A \code{list} contaning the LOO imprecision estimates:
\itemize{
  \item \code{Var_A:} Pooled variance of all sample-variances based on \code{MP_A}
  \item \code{Var_B:} Pooled variance of all sample-variances based on \code{MP_B}
  \item \code{CV_A:} CV estimate based on Var_A and the grand mean of all measurements from \code{MP_A}
  \item \code{CV_B:} CV estimate based on Var_B and the grand mean of all measurements from \code{MP_B}
  \item \code{lambda:} Ratio of pooled variances \code{Var_A} and \code{Var_B}
}
}
\description{
Calculate \code{Var_B}, \code{Var_A}, \code{CV_B}, \code{CV_A} and \code{lambda} without
the \code{loo_id}-th SampleID.
}
\details{
Uses \code{leave_one_out()} to remove one unique value of \code{SampleID} from \code{data}
Then \code{global_imprecision_estimates()} is used to calculate the estimated values
of \code{Var_B}, \code{Var_A}, \code{CV_B}, \code{CV_A} and \code{lambda},
based on the \code{data} without the data from the \code{loo_id}-th \code{SampleID} value.
}
\examples{
library(fasteqa)
# Estimate imprecision estimates without the second clinical sample
gpe <- loo_global_precision_estimates(test_data, 2L)

# Output
print(as.data.frame(gpe))

}
