% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loo_zeta_estimates}
\alias{loo_zeta_estimates}
\title{Calculate the \eqn{\hat{\zeta}} Value When One Observation is Removed}
\arguments{
\item{data}{A \code{list} or \code{data.table}. Must contain
\code{SampleID}, \code{ReplicateID}, \code{MP_A} and \code{MP_B}.}
}
\value{
A \code{list} contaning the resampled \eqn{\hat{\zeta}}.
}
\description{
Resample \code{zeta} one time.
}
\details{
One unique value of \code{SampleID} is removed from \code{data}.
Then \code{estimate_zeta()} is used to calculate \eqn{\hat{\zeta}},
based on \code{data} without the data from the \code{loo_id}-th \code{SampleID} value.
}
\examples{
\dontrun{
  # Estimate zeta without the second clinical sample
  print(loo_zeta_estimatees(test_data, 2L)$zeta)
}
}
