% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MeanOfReplicates}
\alias{MeanOfReplicates}
\title{Calculate means of replicates}
\arguments{
\item{data}{A data frame or data table converted to a list so that every column is unique list elements}

\item{silence}{should progress reports be outputted? For debugging and additional information, where you set \code{silence} to either -1 or 0. Default is 1, which means that no progress reports should be outputted}
}
\value{
A list containing the elements that is needed to build the resulting data of results. Use \code{setDT()} for maximum efficiency when converting the list into a data table
}
\description{
Constructing prediction intervals require EQA data to be on mean-of-replicates-form. So this function should be applied to the EQAM data before it is included into the \code{PredictDeming()} or \code{PredictOLS()}
}
\details{
Calculates means of replicate

The difference between this function and \code{mean_of_replicates()} method in the \code{commutability.selectivity}, is that this is more than ten times faster.
}
\examples{
\dontrun{
  library(commutability.selectivity)
  library(data.table)
  data <- sampled_cs_measurements
  data <- MS_wise(data = data)[Comparison=="MP1 - MP2",]
  data$SampleID <- as.character(data$SampleID)
  data$ReplicateID <- as.character(data$ReplicateID)
  input <- as.list(data)
  output <- setDT(MeanOfReplicates(data = input, silence = 1))
}
}
