% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ols_regression}
\alias{ols_regression}
\title{Fit a Ordinary Least Squares Model}
\usage{
ols_regression(data)
}
\arguments{
\item{data}{A \code{list} or \code{data.table}. Must contain:
\itemize{
    \item \code{MP_A: } A \code{numeric} vector. The means of
    replicated measurements from IVD-MD \code{MP_A} (response).
    \item \code{MP_B: } A \code{numeric} vector. The means of
    replicated measurements from IVD-MD \code{MP_B} (predictor).
}}
}
\value{
A \code{list}. Contains information on the fitted model.
}
\description{
Fits a Ordinary Least Squares (OLS) regression model to \code{data}.
}
\details{
No details are needed. It is OLS.
}
\examples{
# Reproducibility
set.seed(99)

print(rbinom(1, 5, 0.5))

}
