% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pb_test}
\alias{pb_test}
\alias{bp_test}
\title{Perform a Breusch-Pagan Test}
\usage{
bp_test(data, koenker = TRUE)
}
\arguments{
\item{data}{A \code{list} or \code{data.table}. Must contain:
\itemize{
    \item \code{MP_A: } A \code{numeric} vector. The means of
    replicated measurements from IVD-MD \code{MP_A} (response).
    \item \code{MP_B: } A \code{numeric} vector. The means of
    replicated measurements from IVD-MD \code{MP_B} (predictor).
}}

\item{koenker}{A \code{logical} value. If \code{TRUE} (default), the Koenker
modification is applied to the make the test more robust
when data is non-normal.}
}
\value{
A \code{double}. The calculated p-value of the test.
}
\description{
Performs a Breusch Pagan test on variance heterogeneity on \code{data}.
}
\details{
The Breusch-Pagan test and its Koenker modification are both used to detect
heteroskedasticity in linear regression models.

The Koenker version, proposed by Koenker in 1981, modifies the test to make
it more robust when the data is non-normal. Note however, that the Koenker
version will have poorer power than the unmodified Breusch-Pagan test if
the data is close to normal.

Note: \code{NA}-values are silently removed prior to calculating the 
Breusch-Pagan test statistic.
}
\examples{

# Required packages
library(fasteqa)

# Reproducibility
set.seed(99)

# Simulate some example data
x <- runif(n = 50, min = 0, max = 1)
y <- 0.1 + 0.9 * x + rnorm(n = 50, mean = 0, sd = 0.05 * (x + 1))
data <- list(MP_A = y,
             MP_B = x)

# The output (rounded to 3L)
print(round(bp_test(data), 3L))


}
