% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PredictOLS}
\alias{PredictOLS}
\title{Estimation of prediction intervals or prediction band based on OLS regression}
\arguments{
\item{data}{A data table or data frame converted to a list so that each column is a unique list element. Must contain the following columns: MP_A and MP_B}

\item{NewData}{A numeric vector of which x values are used to predict y}

\item{Precision}{A list of the Precision estimates based on \code{CharacterEstimatePrecision()}}

\item{MaxR}{The number of replicates for the majority of clinical samples. For example, if 18 of 25 CSs are meausured in triplicate, and the remaining seven is measured in duplicate, we use MaxR = 3}

\item{level}{the confidence level of the pointwise prediction intervals constructed. If the prediction intervals are used to evaluate commutability of EQAMs, you must remember to adjusts for simultanuous testing, e.g., Bonferroni correction}

\item{silence}{an integer that controls progress reports. Set to 0 for these progress reports to appear. For additional debugging information, set this to \code{-1}. Default is \code{1}, which means that only the output will be returned, and no messages}
}
\value{
A list with four elements, which are, nx (inputted new data), fit (predicted y values based on new data), lwr (lower prediction interval bound), upr (upper prediction interval bound). Use \code{setDT()} from data.table to merge the list elements into a data table
}
\description{
Estimation of prediction intervals based on the OLS regression model is relevant in e.g., commutability evaluation of EQAMs
}
\details{
Estimate pointwise prediction intervals based on OLS regression
}
