% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resample_fun_data}
\alias{resample_fun_data}
\title{Resample fun-of-replicates data}
\arguments{
\item{data}{\code{list} or \code{data table} - Must contain \code{SampleID}, \code{MP_A} and \code{MP_B}. \code{SampleID} and \code{ReplicateID} must be of character type, or else an error is thrown}

\item{make_unique}{- \code{integer} that controls the output SampleID. If \code{make_unique = 1}, the default, new SampleIDs will be made. Conversely \code{make_unique = 0} will use the original SampleIDs. The latter is not recommended because potential calculations are affected by this}
}
\value{
A list containing the resampled fun-of-replicates data. Use \code{setDT()} for maximum efficiency if you desire to convert the resampled data to a data table
}
\description{
In order to construct bootstrap confidence intervals and do inference on a set of population parameters where the underlying distribution is complex, will require bootstrap replicates. This function is both efficient and does its job, but at a cost of a strict input requirement
}
\details{
Resample fun-of-replicates data

Combine with e.g., \code{predict_eqa()}, to estimate classification rates and more
}
\examples{
\dontrun{
  library(fasteqa)
  data <- simulate_data_eqa(list(n = 25, R = 3, cvx = 0.06, cvy = 0.04))
  data$SampleID <- as.character(data$SampleID)
  data$ReplicateID <- as.character(data$ReplicateID)
  data <- fun_of_replicates(data, fun = "median")
  resampled_data <- resample_fun_data(data)
}
}
