% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resample_fun_of_samples}
\alias{resample_fun_of_samples}
\title{Resample Cluster Statistics Based on External Quality Assessment (EQA) Clinical Sample Data}
\arguments{
\item{data}{A \code{list} or \code{data.table}. The mean of replicates (MOR)
clinical sample data. Must contain:
\itemize{
   \item \code{SampleID}: A \code{character} vector of sample identifiers.
   \item \code{MP_A}: A \code{numeric} vector of measurements from method A (response).
   \item \code{MP_B}: A \code{numeric} vector of measurements from method B (predictor).
}}
}
\value{
A \code{list} that contains the resampled \code{data}.
}
\description{
This function resamples resamples sample statistics based on EQA clinical sample data
with replacement. This function can for example be used to create bootstrap-based estimations
of predictions and conclusion strengths.
}
\details{
This function is an optimized method to resample clinical sample statistics
based on clinical sample data.

Note, if \code{ReplicateID} is part of \code{data}, this method may yield
unexpected results. For methods that require resampling of all replicated
measurements, please use \code{resample_samples()} instead.

Note also that the \code{SampleID} of the returned object does not correspond
with the actual values in the input \code{SampleID}. Thus, this function
should not be used to do inference on individual clinical samples!
}
\examples{
library(fasteqa)
mor_test_data <- fun_of_replicates(test_data)
mor_test_data$MP_A <- round(mor_test_data$MP_A, 2)
mor_test_data$MP_B <- round(mor_test_data$MP_B, 2)
resampled_data <- resample_fun_of_samples(mor_test_data)
print(as.data.frame(resampled_data))
}
