% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resample_global_precision_estimates}
\alias{resample_global_precision_estimates}
\title{Resample imprecision estimates based on clustered EQA clinical sample data}
\arguments{
\item{data}{A \code{list} or \code{data.table}. Must contain
\code{SampleID}, \code{ReplicateID}, \code{MP_A} and \code{MP_B}.}
}
\value{
A \code{list} contaning the resampled imprecision estimates:
\itemize{
  \item \code{Var_A:} Pooled variance of all sample-variances based on \code{MP_A}
  \item \code{Var_B:} Pooled variance of all sample-variances based on \code{MP_B}
  \item \code{CV_A:} CV estimate based on Var_A and the grand mean of all measurements from \code{MP_A}
  \item \code{CV_B:} CV estimate based on Var_B and the grand mean of all measurements from \code{MP_B}
  \item \code{lambda:} Ratio of pooled variances \code{Var_A} and \code{Var_B}
}
}
\description{
Resample \code{Var_B}, \code{Var_A}, \code{CV_B}, \code{CV_A} and \code{lambda} one time.
}
\details{
Uses \code{resample_samples()} to resample \code{data} on sample-level.
Then \code{global_imprecision_estimates()} is used to calculate the estimated values
of \code{Var_B}, \code{Var_A}, \code{CV_B}, \code{CV_A} and \code{lambda},
based on the resampled \code{data}.
}
\examples{
library(fasteqa)

# Resample global imprecision estimates 3 times
rie <- replicate(n = 3,
                 expr = resample_global_precision_estimates(test_data),
                 simplify = FALSE)
rie <- lapply(X = rie,
              FUN = as.data.frame)
resampled_impr_estimates <- rbind(rie[[1]],
                                  rie[[2]],
                                  rie[[3]])
print(resampled_impr_estimates)

}
