% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resample_zeta_estimates}
\alias{resample_zeta_estimates}
\title{Resample \eqn{\hat{\zeta}} values based on clustered EQA clinical sample data}
\arguments{
\item{data}{A \code{list} or \code{data.table}. Must contain
\code{SampleID}, \code{ReplicateID}, \code{MP_A} and \code{MP_B}.}
}
\value{
A \code{list} contaning the resampled \eqn{\hat{\zeta}}.
}
\description{
Resample \code{zeta} one time.
}
\details{
Uses \code{resample_samples()} to resample \code{data} on sample-level.
Then \code{estimate_zeta()} is used to calculate \eqn{\hat{\zeta}},
based on the resampled \code{data}.
}
\examples{
# Resample global imprecision estimates 1000 times
rze <- replicate(n = 1000,
                 expr = resample_zeta_estimates(test_data)$zeta,
                 simplify = TRUE)
# Get summary statistics of resampled zeta estimates
summary(rze)

}
