% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{residuals_eqa}
\alias{residuals_eqa}
\title{Obtain the residuals based on data and imprecision estimates}
\arguments{
\item{data}{A \code{list} or \code{data.table} that includes mean values of replicated clinical sample data, grouped by \code{SampleID}. The data for the two compared IVD-MDs under consideration should be found in the columns named \code{MP_A} and \code{MP_B}. The \code{SampleID} should be a character vector, whereas \code{MP_A} and \code{MP_B} must be numeric vectors.}

\item{imprecision_estimates}{A \code{list} or \code{data.table} carrying imprecision estimates for IVD-MD, ideally matching those provided by the \code{global_precision_estimates()} function. A non-missing positive value for \code{lambda} is the minimum requirement.}

\item{method}{A \code{character} string specifying the linear regression model applied for obtaining the residuals and fitted values. The default value is \code{fg}. The possible values for \code{method} include:
\itemize{
  \item{\code{fg: }}{Refers to Deming regression, incorporating elements from the work of W. Fuller and J. Gillard.}
  \item{\code{clsi: }}{Refers to Deming regression, based on the CLSI EP14 standard.}
  \item{\code{ols: }}{Represents ordinary least squares regression.}
}}

\item{studentize}{An \code{integer}. When set to \code{1} (default), the residuals get studentized, implying the mean of the residuals is subtracted from each residual and then divided by the standard deviation of the residuals. If set to \code{0}, residuals are not studentized. Studentized residuals can be beneficial if there is a need to compare the residuals' distribution to the standard normal distribution or a specific t-distribution.}
}
\value{
This function returns a \code{list} comprising two numeric vectors named \code{residuals} and \code{fitted}. These are the residuals and fitted values, respectively, based on the input \code{data}, \code{imprecision_estimates}, and \code{method}.
}
\description{
This function efficiently extracts the residuals and fitted values based on the provided \code{data}, \code{imprecision_estimates}, and \code{method}.
}
\details{
Obtain the residuals based on data and imprecision estimates

While the outcome differences between the three provided methods may be minute, they can still be noteworthy. Suppose 'n' is the count of unique clinical samples. The \code{fg} method estimates latent concentration values and utilizes n - 1 degrees of freedom. In contrast, \code{clsi} does not estimate latent variables and uses 'n' degrees of freedom. The model associated with \code{ols} overlooks the imprecision in \code{MP_B}.
}
\examples{
\dontrun{
  library(fasteqa)
  parameters <- list(n = 20, R = 3, cvx = 0.03, cvy = 0.02, cil = 10, ciu = 15)
  data <- simulate_eqa_data(training_parameters)
  data$SampleID <- as.character(data$SampleID)
  data$ReplicateID <- as.character(data$ReplicateID)
  imprecision <- global_precision_estimates(data)
  mean_of_replicates_data <- fun_of_replicates(data)
  # Extracting raw residuals based on the clsi Deming regression approach
  residuals_eqa(mean_of_replicates_data, imprecision, "clsi", 0)
}
}
