% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_eqa_data}
\alias{sim_eqa_data}
\title{Simulation of External Quality Assessment (EQA) Data}
\arguments{
\item{parameters}{A \code{list} of parameters and their value. These values
       are used to simulate the EQA data. Optional parameters that you may
       include into the list are:
\itemize{
  \item \code{n: } The number of samples.
  \item \code{R: } The number of replicates on each sample.
  \item \code{cvx: } The repeatability coefficient of variation for IVD-MD 
                     \code{MP_B}.
  \item \code{cvy: } The repeatability coefficient of variation for IVD-MD
                     \code{MP_A}.
  \item \code{cil: } The lower bound of the concentration interval.
  \item \code{ciu: } The upper bound of the concentration interval.
  \item \code{dist: } The distribution to simulate latent variables from.
                      Possbile choices include
                      \code{unif} (uniform distribution, default),
                      \code{norm} (normal distribution),
                      \code{lst} (location-scale student t-distribution),
                      \code{lnorm} (log-normal distribution).
  \item \code{df_tau: } The degrees of freedom for the \code{lst}
                        distribution if the distribution of latent variables
                        are location-scale student t-distributed ('lst').
                        Defaults to \eqn{5} if not otherwise is specified.
  \item \code{eta: } The heteroscedasticity factor \eqn{\eta}.
  \item \code{eta0: } The proportion of "base standard deviation". See
                      details.
  \item \code{qpos: } Position of systematic differences in non-selectivity.
                      \code{0} signify lower concentration interval and
                      \code{1} signify upper.
  \item \code{qran: } Interquantile range where systematic differences in
                      non-selectivity should have its effect
  \item \code{prop: } average proportion of clinical samples affected by
                      random differences in non-selectivity.
  \item \code{mmax: } The maximum relocation magnitude multiplier,
                      \eqn{m_{\max}}. This number is a multiplier of the
                      base standard deviation. This will only have an effect
                      if \code{prop}, or \code{qpos} and \code{qran} are
                      specified as well.
  \item \code{b0: } For systematic linear bias between IVD-MDs.
                    The intercept. Defaults to 0.
  \item \code{b1: } For systematic linear bias between IVD-MDs. The 
                    slope. Defaults to 1.
  \item \code{c0: } For systematic linear non-selectivity in IVD-MD 1.
                    The intercept. Defaults to 0.
  \item \code{c1: } For systematic linear non-selectivity in IVD-MD 1.
                    The slope. Defaults to 1.
  \item \code{error_dist: } The distribution to simulate measurement error
                            components from. Possible choices include 'norm'
                            (normal distribution, default) and
                            'lt' (location student t-distribution).
  \item \code{dfx: } The degrees of freedom for the measurement error
                     components in IVD-MD 1 if \code{error_dist = 'lt'}.
                     Defaults to \eqn{5} if not otherwise is specified.
  \item \code{dfy: } The degrees of freedom for the measurement error
                     components in IVD-MD 2 if \code{error_dist = 'lt'}.
                     Defaults to \eqn{5} if not otherwise is specified.
  \item \code{md_method: } Method for simulating missing data. Possible
                           choices include \code{'none'}
                           (no missing data is simulated, default),
                           \code{'mar'} (missing at random),
                           \code{'mnar'} (missing not at random) and
                           \code{'marmnar'}
                           (missing both at random and not at random)
  \item \code{mar_prob: } The probability (value between 0 and 1) of having
                          a missing measurement. Only relevant if
                          \code{md_method} is \code{'mar'} or
                          \code{'marmnar'}. If not specified, but 
                          \code{md_method = 'mar'} or
                          \code{md_method = 'marmnar'}, it defaults to 0.05.
  \item \code{mnar_threshold: } The lower bound threshold (\code{double})
                                for when a measurement should be missing.
                                Only relevant if \code{md_method} is
                                \code{'mnar'} or \code{'marmnar'}. If not
                                specified, but \code{md_method = 'mnar'}
                                or \code{md_method = 'marmnar'}, it defaults
                                to \code{cil}. Alternatively, if not
                                specified, but \code{md_method = 'mnar0'}
                                or \code{md_method = 'marmnar0'},
                                it defaults to 0.
  \item \code{g: } A custom function to simulate \eqn{\mathrm{E}[y_{ir}]}
                   from. It is not recommended to use this for linear
                   \code{g}, because it is much slower. This is generally
                   used to simulate data from non-linear functions other
                   than the three built-in alternatives
                   (see \code{type} argument).
  \item \code{obs_tau: } A \code{numeric} vector of \eqn{\tau_i} values.
                         (true but latent values for IVD-MD one).
                         If this parameter is specified, \code{n} will take
                         the length of \code{obs_tau}. Moreover, all
                         parameters used to generate \eqn{\tau_i} are
                         ignored.
}}

\item{type}{A \code{integer}. Custom built-in functions \code{g}. Set to
\code{0}, to use a linear \code{g}. Otherwise, particular custom
non-linear functions can be used. The four different built-in
\code{g}:
\itemize{
   \item \code{type = 0: } \eqn{g(\tau_i) = \beta_0 + \beta_1 \cdot
                                \tau_i}
   \item \code{type = 1: } \eqn{g(\tau_i) = \tau_i + 0.90 \cdot
                                \mathrm{sin}(0.40 \cdot
                                \tau_i ^ {1.06})}
   \item \code{type = 2: } \eqn{g(\tau_i) = \tau_i + 0.05 \cdot
                                \exp(0.16 \cdot \tau_i ^ {1.35})}
   \item \code{type = 3: } \eqn{g(\tau_i) = \tau_i - \exp[
                                -0.125 \cdot (\tau_i - 1.50)^2]}
}
If \code{g} exists in \code{parameters}, this argument is ignored.}

\item{AR}{A \code{logical} value. If \code{TRUE}, data is simulated including
replicated measurements. Otherwise, mean of replicated measurements
are returned (MOR).}

\item{include_parameters}{A \code{logical} value. If \code{TRUE}, the used
parameters in the data simulation is saved and placed in a seperate
list as part of the output.}
}
\value{
A \code{list} of length three or four:
\itemize{
  \item \code{SampleID:} An \code{integer} vector. Sample identifiers.
  \item \code{ReplicateID:} An \code{integer} vector. Replicate identifiers.
                            Only appears if \code{AR = TRUE}.
  \item \code{MP_A:} An \code{numeric} vector. Simulated measurements from
                     one of the IVD-MDs in comparison.
  \item \code{MP_B:} An \code{numeric} vector. Simulated measurements from
                     the other IVD-MD in comparison.
}
}
\description{
Simulates a dataset with at most n x R rows measurements.
}
\details{
Simulates method comparison data for \code{n} samples
(e.g., clinical samples, pools, external quality assessment samples, reference material samples),
where each sample is measured at most \code{R} times (replicated measurements).
In theory, we simulate measurement pairs \eqn{(x_{ir}, y_{ir})} from:

\eqn{x_{ir} = f(\xi_i) + h_{ir} \newline
     y_{ir} = g(f(\xi_i)) + v_{ir} \newline
     g(f(\xi_i)) = f(\xi_i) + e_{i}}.

If \eqn{f} is linear in \eqn{\xi_i}, we have that
\eqn{\tau_i = f(\xi_i) = c_0 + c_1 \, \xi_i}.
We generally do not care much about the shape of \eqn{f(\xi_i)}. We are
more interested in the values \eqn{\tau_i} and the shape of \eqn{g}.
Assuming that \eqn{g} is linear, we have that

\eqn{g(\tau_i) = \beta_0 + \beta_1 \cdot \tau_i + e_{i}}.

Thus, we simulate \eqn{(x_{ir}, y_{ir})} from:

\eqn{x_{ir} = \tau_i + h_{ir} \newline
     y_{ir} = g(\tau_i) + v_{ir}},

where \eqn{g} is some function.

If we wish to specify \eqn{f} (not very informative) we can specify it
through \code{c0} and \code{c1}. \eqn{g} is specified through numerous
parameters such as \code{b0}, \code{b1}, \code{qpos}, \code{qran},
\code{mmax}, \code{prop} and \code{type}. One can also specify \eqn{g}
using a custom R function. See explanation of \code{parameters}.

\eqn{\tau_i} is a random variable modelled through \code{cil}, \code{ciu},
\code{dist} and \code{df_tau}.

\eqn{h_{ir}} and \eqn{v_{ir}} are the measurement error components and
are modelled through \code{cvx}, \code{cvy}. The coefficients of variation,
\code{cvx}, \code{cvy} can be functions of \code{error_dist}, \code{dfx},
\code{dfy}, \code{eta} and \code{eta0}.

In order to convert the outputted \code{list} to a data frame structure,
use either \code{as.data.frame()}, \code{as.data.table()} or
\code{as.tibble()}.  The most efficient way to convert the output is
however \code{setDT()} from the \code{data.table} package.

Note, there is no guarantee that every simulated \eqn{(x_{ir}, y_{ir})}
will take positive values. To avoid negative values, you will need to
select values of \code{cil}, \code{ciu}, \code{cvx}, \code{cvy}, and
\code{mmax} that result in a negligble probability of generating negative
values. In general, avoid using large values of \code{cvx} and \code{cvy}
if \code{cil} is close to zero and \code{ciu} - \code{ciu} is large.
Negative values can alternatively be replaced with \code{NA} values if
\code{md_method = 'mnar0'}.
}
\examples{
# Load package
library(fasteqa)

# Simulate 25 clinical samples measure in triplicate affected by random
# differences in non-selectivity

parameters_cs <- list(n = 25,
                      R = 3,
                      cvx = 0.01,
                      cvy = 0.01,
                      cil = 2,
                      ciu = 10,
                      prop = 0.2,
                      mmax = 5)
simulated_cs <- sim_eqa_data(parameters_cs,
                             type = 0,
                             AR = TRUE)
print(as.data.frame(simulated_cs))


# Simulate 30 clinical samples and 3 External Quality Assessment material
# samples (all measured in duplicate) that are commutable with the
# clinical samples.

parameters_cs <- list(n = 30,
                      R = 2,
                      cvx = 0.01,
                      cvy = 0.01,
                      cil = 5,
                      ciu = 90,
                      qran = 0.30,
                      qpos = 1,
                      mmax = 7.5)

simulated_cs <- sim_eqa_data(parameters_cs,
                             type = 0,
                             AR = TRUE,
                             include_parameters = TRUE)
true_values_eqa_samples <- c(10, 40, 75)                             
parameters_eq <- simulated_cs$parameters
parameters_eq <- c(list(obs_tau = true_values_eqa_samples),
                   parameters_eq)
simulated_cs <- simulated_cs$simulated_data
simulated_eq <- sim_eqa_data(parameters_eq,
                             type = 0,
                             AR = TRUE)
print(as.data.frame(simulated_cs))
print(as.data.frame(simulated_eq))
                             

}
