% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skewness}
\alias{skewness}
\title{Calculate Sample Skewness of a Random Sample}
\arguments{
\item{x}{A \code{numeric} vector. A random sample.}

\item{na_rm}{A \code{logical} value. If \code{TRUE}, \code{NA}-values are
removed prior to calculation of the sample skewness.}
}
\value{
A \code{double}. The calculated sample skewness.
}
\description{
Calculates the sample skewness of a random sample \code{x}, using the
adjusted Fisher-Pearson standardized moment coefficient.
}
\details{
Calculates the sample skewness of a random sample
\eqn{\lbrace x_i \rbrace_{i=1}^{N}}. The sample skewness, attempts to
estimate the theoretical skewness, \eqn{\gamma}, by using
the following estimator (adjusted Fisher-Pearson standardized moment
coefficient):

\eqn{\hat{\gamma} = \frac{\sqrt{N(N-1)}}{N-2} \cdot
\frac{\frac{1}{N}\sum_{i=1}^{N}(x_i - \overline{x})^3}
{\Big[\frac{1}{N}\sum_{i=1}^{N}(x_i - \overline{x})^2\Big]^{1.5}}}.

Note: This estimator biased if not \eqn{x_i \sim \mathrm{N}(\mu, \sigma^2)}.
}
\examples{

# Reproducibility
set.seed(99)

# Simulate 1 million observations from log-normal(0, 0.25)
y <- rlnorm(n = 1e6, meanlog = 0, sdlog = 0.25)
sample_skew <- skewness(y)

# Theoretical skew for log-normal(0, 0.25)
theoretical_skew <- (exp(0.25^2) + 2) * sqrt(exp(0.25^2) - 1)

cat("Theoretical skewness:", round(theoretical_skew), "\n",
    " Empirical skewness:", round(sample_skew), "\n")


}
