% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Fictive Clinical Sample Data}
\format{
## `test_data`
A data frame with 60 rows and 4 columns:
\describe{
  \item{SampleID}{ID of clinical sample}
  \item{ReplicateID}{ID of replicated measurements}
  \item{MP_A}{First IVD-MD in the comparison}
  \item{MP_B}{Second IVD-MD in the comparison}
}
}
\source{
clsi
}
\usage{
test_data
}
\description{
Fictive Clinical Sample Data
}
\keyword{datasets}
