% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BCa_bootstrap_ci}
\alias{BCa_bootstrap_ci}
\title{Estimate bias and skewness corrected bootstrap confidence intervals}
\arguments{
\item{bootstrapped_parameter_estimates}{Numeric vector - Bootstrapped estimates of the parameter of interest}

\item{jackknife_parameter_estimates}{Numeric vector - Jack knife estimates of the parameter of interest}

\item{original_parameter_estimate}{Double - Original point estimate of the parameter of interest}

\item{level}{Double - Required confidence level for the estimated confidence interval}

\item{silence}{Integer - Controls the progress reports outputted for debugging and further examination of the command. \code{silence = -1} or \code{silence = 0} signify that progress reports should be printed. Default is \code{silence = 1} which suppresses all printing}
}
\value{
A numeric vector. The first element of the vector is the lower bound of the confidence interval, and the second element of the vector is the upper bound of the confidence interval
}
\description{
Estimation of bootstrap confidence intervals for a parameter of interest using the BCa-variant
}
\details{
Estimate bias and skewness corrected bootstrap confidence intervals

To get an adequate estimate of the theoretical confidence interval, it is essential that original data represents the underlying distribution for the population parameter. More data typically corresponds with larger probability of our data being representative. For bootstrap confidence interval estimation independent of jack knife estimates, \code{bootstrap_ci()} may be used instead
}
\examples{
\dontrun{
  fictional_original_estimate <- 2.1
  fictional_bootstrap_estimates <- runif(n = 1e3, min = 0.5, max = 2.5)
  fictional_jackknife_estimates <- runif(n = 1e1, min = 1.5, max = 2.5)
  BCa_bootstrap_ci(fictional_bootstrap_estimates,
                   fictional_jackknife_estimates,
                   fictional_original_estimate,
                   level = 0.95,
                   silence = 1)
}
}
