% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bootstrap_ci}
\alias{bootstrap_ci}
\title{Estimate simple bootstrap confidence intervals}
\arguments{
\item{bootstrapped_parameter_estimates}{A \code{numeric vector} storing bootstrapped estimates of the target parameter. A couple of missing values are allowed and will be removed if present.}

\item{original_parameter_estimate}{A \code{double} representing the original point estimate of the target parameter. A missing value is not allowed.}

\item{type}{An \code{integer} that refers to the type of bootstrap confidence interval to be estimated. Valid inputs include:
\itemize{
  \item{\code{1: }}{Standard normal bootstrap confidence interval}
  \item{\code{2: }}{Basic bootstrap confidence interval}
  \item{\code{3: }}{Percentile bootstrap confidence interval}
}}

\item{level}{A \code{double} between 0 and 1, indicating the desired confidence level for the estimated bootstrap confidence interval of the target parameter. Defaulting to \code{0.95}.}

\item{silence}{An \code{integer} controlling the level of progress reporting for debugging and analysis. Set to \code{-1} or \code{0} for detailed reports. Defaults to \code{1} to suppress output.}
}
\value{
A \code{numeric vector} with two entries: the lower and upper bounds of the bootstrap confidence interval, respectively.
}
\description{
Estimation of bootstrap confidence intervals for a parameter of interest based on \code{bootstrapped_parameter_estimates}, \code{original_parameter_estimate}, \code{type} and \code{level}.
}
\details{
Estimate simple bootstrap confidence intervals

To get an adequate estimate of the theoretical confidence interval for the target parameter, it is essential that the original data represents the underlying distribution for the population parameter. More data typically corresponds with larger probability of our data being representative. For second-order accurate and transformation respective confidence intervals, you may employ \code{BCa_bootstrap_ci()} instead. Note that \code{BCa_bootstrap_ci()} requires jack knife estimates of the parameter of interest as well as bootstrap estimates
}
\examples{
\dontrun{
  fictional_original_estimate <- 2.1
  fictional_bootstrap_estimates <- runif(n = 1e3, min = 0.5, max = 2.5)
  bootstrap_ci(fictional_bootstrap_estimates,
               fictional_original_estimate,
               type = 3,
               level = 0.95,
               silence = 1)
}
}
