% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bootstrap_ci}
\alias{bootstrap_ci}
\title{Estimate simple bootstrap confidence intervals}
\arguments{
\item{bootstrapped_parameter_estimates}{Numeric vector - Bootstrapped estimates of the parameter of interest}

\item{original_parameter_estimate}{Double - Original point estimate of the parameter of interest}

\item{type}{Integer - Which type of bootstrap confidence interval is required
\itemize{
  \item{\code{1: }}{Standard normal bootstrap confidence interval}
  \item{\code{2: }}{Basic bootstrap confidence interval}
  \item{\code{3: }}{Percentile bootstrap confidence interval}
}}

\item{level}{Double - Required confidence level for the estimated confidence interval}

\item{silence}{Integer - Controls the progress reports outputted for debugging and further examination of the command. \code{silence = -1} or \code{silence = 0} signify that progress reports should be printed. Default is \code{silence = 1} which suppresses all printing}
}
\value{
A numeric vector. The first element of the vector is the lower bound of the confidence interval, and the second element of the vector is the upper bound of the confidence interval
}
\description{
Estimation of bootstrap confidence intervals for a parameter of interest
}
\details{
Estimate simple bootstrap confidence intervals

To get an adequate estimate of the theoretical confidence interval, it is essential that original data represents the underlying distribution for the population parameter. More data typically corresponds with larger probability of our data being representative. For second-order accurate and transformation respective confidence intervals, you may employ \code{BCa_bootstrap_ci()} instead. Note that \code{BCa_bootstrap_ci()} require jack knife estimates of the parameter of interest as well as bootstrap estimates
}
\examples{
\dontrun{
  fictional_original_estimate <- 2.1
  fictional_bootstrap_estimates <- runif(n = 1e3, min = 0.5, max = 2.5)
  bootstrap_ci(fictional_bootstrap_estimates,
               fictional_original_estimate,
               type = 3,
               level = 0.95,
               silence = 1)
}
}
