% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{global_precision_estimates}
\alias{global_precision_estimates}
\title{Calculate imprecision point estimates of measurements in a given MS comparison}
\arguments{
\item{data}{\code{list} or \code{data table} - Data with elements/columns \code{SampleID}, \code{ReplicateID}, \code{MP_A} and \code{MP_B}}

\item{silence}{\code{integer} - Should debugging messages be printed. Default is no}
}
\value{
\code{list} - with point imprecision estimates \code{Var_A}, \code{Var_B}, \code{CV_A}, \code{CV_B} and \code{lambda}
}
\description{
Structure-requirement of \code{data}:
\itemize{
  \item{\code{SampleID: }}{Must be a character vector}
  \item{\code{ReplicateID: }}{Must be a character vector}
  \item{\code{MP_A: }}{Must be a numeric vector}
  \item{\code{MP_B: }}{Must be a numeric vector}
}
}
\details{
Calculate imprecision point estimates of measurements in a given MS comparison

Calculates various global imprecision estimates. To get CVs in percent you need only to multiply the raw CV estimates with 100. Here is a rough explaination of the output list:
\itemize{
  \item{\code{Var_A: }}{Pooled variance of all sample-variances based on MS_A}
  \item{\code{Var_B: }}{Pooled variance of all sample-variances based on MS_B}
  \item{\code{CV_A: }}{Global CV estimate based on Var_A and the grand mean of all measurements from MS_A}
  \item{\code{CV_B: }}{Global CV estimate based on Var_B and the grand mean of all measurements from MS_B}
  \item{\code{lambda: }}{Ratio of pooled variances Var_A and Var_B}
}
}
\examples{
\dontrun{
  library(fasteqa)
  data <- simulate_eqa_data(list(n = 25, R = 3, cvx = 0.02, cvy = 0.3))
  data$SampleID <- as.character(data$SampleID)
  data$ReplicateID <- as.character(data$ReplicateID)
  global_prcision_estimates(data = data)
}
}
