% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{leave_one_out}
\alias{leave_one_out}
\title{Leave-one-out on clustered EQA clinical sample data}
\arguments{
\item{data}{\code{List} or \code{data table} - Data with list elements or data table columns with names \code{SampleID}, \code{ReplicateID}, \code{MP_A} and \code{MP_B}. \code{SampleID} and \code{ReplicateID} must be of character type for the function to operate correctly}

\item{loo_id}{Integer - Which of the samples of \code{SampleID} should be left out. Default value is 1}
}
\value{
A \code{list} containing the original data, but without the sample id corresponding to the given \code{loo_id}
}
\description{
Needed to calculate jack knife estimates of a parameter, that is required when using \code{BCa_bootstrap_ci()}. Alternatively one could use the Jack knife estimates to calculate standard error or bias of the estimator of relevance
}
\details{
Leave-one-out on clustered EQA clinical sample data

loo_ids can not be vectorized directly in R. Use \code{sapply()} or \code{replicate()} to leave out sample IDs one by one
}
\examples{
\dontrun{
  library(commutability.selectivity)
  data <- sdwdnsp2()
  loo_data <- leave_one_out(data, loo_id = 5)
}
}
