% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_results}
\alias{merge_results}
\title{Merge all computations efficiently into one object}
\arguments{
\item{pb_data}{\code{list} or \code{data table} - Prediction band data, which must at least contain \code{comparison}, \code{predictor}, \code{prediction}, \code{lwr} and \code{upr}}

\item{ce_data}{\code{list} or \code{data table} - Commutability evaluation data for evaluated material, which must at least contain \code{comparison}, \code{SampleID}, \code{MP_B}, \code{MP_A}, \code{prediction}, \code{lwr}, \code{upr} and \code{inside}}

\item{zeta_data}{\code{list} or \code{data table} - Zeta estimates, which must at least contain \code{comparison}, \code{zeta}, \code{lwr}, \code{upr}, \code{zeta_critical}, \code{zeta_conclusion}}

\item{imprecision_data}{\code{list} or \code{data table} - global imprecision estimates, which must at least contain \code{comparison}, \code{CV_A}, \code{CV_A_lwr}, \code{CV_A_upr}, \code{CV_B}, \code{CV_B_lwr}, \code{CV_B_upr}, \code{lambda}, \code{lambda_lwr} and \code{lambda_upr}. This one may be omitted, if one is uninterested in viewing the imprecision estimates}

\item{rounding}{\code{integer} - How many decimals should be included in the float type variables? The implicit maximum is 6 decimals, meaning that any integer larger than 6 will produce a warning}

\item{include_imprecision_estimates}{\code{boolean} - Should imprecision estimates be part of the merged results. Default is \code{FALSE}.}

\item{silence}{\code{integer} - Should progress reports be printed? For yes, pass \code{0}. For no, pass \code{1}. \code{1} is default.}
}
\value{
A \code{list} with two main components, merged data results for commutability evaluation of control materials and merged results for prediction bands. The latter may be used to plot commutability evaluation plots
}
\description{
Merge all the essential components (e.g., pb_data, ce_data, zeta_data) of the commutability evaluation analysis into one object, so it will be easier to plot and present commutability evaluation results
}
\details{
Merge all commutability evaluation computations efficiently into one object

The merging is done with respect to the comparison column that needs to be common in every data input. The effects of missing values in the comparison variable may yield unexpected results or errors, so ensure that at least this column, is NA-free.
}
\examples{
\dontrun{
  library(fasteqa)
  data <- simulate_data_eqa(list(n = 25, R = 3, cvx = 0.06, cvy = 0.04))
  estimate_zeta(data)
}

}
