% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_eqa}
\alias{predict_eqa}
\title{Estimate prediction intervals for EQA data with Deming or OLS}
\arguments{
\item{data}{\code{list} or \code{data table} - Mean-of-replicates clinical sample data with elements/columns \code{SampleID}, \code{MP_A} and \code{MP_B}}

\item{new_data}{\code{list} or \code{data table} - Mean-of-replicates control material data with \code{MP_B}. \code{SampleID} and \code{MP_A} may also be in new_data based on desired output structure:
\itemize{
  \item{\code{Minimum requirement: }}{Must contain \code{MP_B}}
  \item{\code{method = ols: }}{Must at least contain both \code{MP_B} and \code{MP_A}}
  \item{\code{Inside checks performed: }}{Must at least contain both \code{MP_B} and \code{MP_A}}
  \item{\code{Sample-wise prediction}}{Must at least contain \code{MP_B} and \code{SampleID} (plus \code{MP_A} if method = ols)}
}}

\item{imprecision_estimates}{\code{list} or \code{data table} - Imprecision estimates e.g., that which are outputted by \code{global_precision_estimates()}. Minimum requirement: \code{lambda} and \code{Var_B} must be part of the}

\item{R}{integer - Number of replicates, which new_data is based on}

\item{method}{string - Which method should be used to estimate the prediction intervals. Default is \code{fg}. At the moment, possible values for method is
\itemize{
  \item{\code{fg: }}{Standard Deming regression, but prediction intervals are calculated using components from J. Gillards work}
  \item{\code{clsi: }}{Standard Deming regression, but prediction intervals are calculated based on derivation of Jeff Vaks}
  \item{\code{ols: }}{Ordinary least squares regression, but the predictor is chosen so that neglected prediction error variance is reduced}
}}

\item{level}{float - Confidence level of prediction intervals. Should ideally be corrected for simulations testing. Default level is 0.99}

\item{rounding}{integer - How many decimals should be included in the predictions and prediction intervals. Two or three decimals are often sufficient. Default is 3}
}
\value{
list - prediction interval data based on use-inputs
}
\description{
A rich function that calculates prediction intervals for EQA data
}
\details{
Estimate prediction intervals for EQA data with Deming or OLS

If possible it is wise to always include \code{SampleID}, \code{MP_A} and \code{MP_B} in new_data. If by some reason only \code{MP_B} is available, one cannot use \code{method = ols}. For example, when construction predicton bands, we will only have \code{MP_B} is available.
}
\examples{
\dontrun{
  library(fasteqa)
  training_parameters <- list(n = 25, R = 3, cvx = 0.01, cvy = 0.015, cil = 10, ciu = 70)
  test_parameters <- list(n = 5, R = 3, cvx = 0.01, cvy = 0.015, cil = 10, ciu = 70)
  training_data <- simulate_eqa_data(training_parameters)
  test_data <- simulate_eqa_data(test_parameters)
  training_data$$SampleID <- as.character(training_data$SampleID)
  training_data$$ReplicateID <- as.character(training_data$ReplicateID)
  imprecision <- global_precision_estimates(data)
  mean_of_replicates_training_data <- fun_of_replicates(training_data)
  mean_of_replicates_test_data <- fun_of_replicates(test_data)
  prediction_intervals <- predict_eqa(mean_of_replicates_training data,
                                      mean_of_replicates_test_data,
                                      imprecision)
}
}
