% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_eqa}
\alias{predict_eqa}
\title{Prediction Interval Estimation for EQA Data via Deming or OLS Regression}
\arguments{
\item{data}{A \code{list} or \code{data.table} representing mean-of-replicates clinical sample data with the \code{list} elements or \code{data.table} columns: \code{SampleID}, \code{MP_A}, and \code{MP_B}.}

\item{new_data}{A \code{list} or \code{data.table} representing mean-of-replicates EQA material / reference material data. This should at least include \code{MP_B} but may also contain \code{SampleID} and \code{MP_A} depending on the desired output structure:
\itemize{
  \item{\code{Minimum requirement}: }{new_data must include \code{MP_B}}
  \item{\code{method = 'ols'}: }{new_data must include both \code{MP_B} and \code{MP_A}}
  \item{\code{Inside checks performed}: }{new_data must include both \code{MP_B} and \code{MP_A}}
  \item{\code{Sample-wise prediction}: }{new_data must include \code{MP_B} and \code{SampleID} (plus \code{MP_A} if method = 'ols')}
}}

\item{imprecision_estimates}{A \code{list} or \code{data.table} that includes necessary imprecision estimates. If \code{method} is \code{'fg'} or \code{'ols'}, this parameter should contain \code{lambda}. In case of \code{method = 'clsi'}, both \code{lambda} and \code{Var_B} must be included. The function \code{global_precision_estimates()} can be employed to generate suitable input for this parameter directly.}

\item{R}{An \code{integer} indicating the number of replicates on which new_data is based. The convenience function \code{count_samplewise_replicates()} can be employed to generate suitable input for this parameter directly}

\item{R_ratio}{An \code{float} indicating the ratio of replicates between that number data and that number new_data are based on. Only relevant if \code{method = 'fg'} and if the number of replicates of data and new_data differs.}

\item{method}{A \code{string} specifying the method for estimating the prediction intervals. Default is \code{'fg'}. Current possible prediction interval estimation methods are:
\itemize{
  \item{\code{fg: }}{Implements standard Deming regression with prediction intervals calculated using the Fuller and Gillard methods.}
  \item{\code{clsi: }}{Utilizes standard Deming regression, with prediction intervals derived from the CLSI method.}
  \item{\code{ols: }}{Implements Ordinary Least Squares regression, where the predictor is selected to minimize the variance of the ignored IVD-MD uncertainty. Specifically, \code{MP_A} is used as the predictor when \code{lambda < 1}, otherwise, \code{MP_B} is utilized.}
}}

\item{level}{A \code{float} representing the confidence level for the prediction intervals. It should be between \code{0} and \code{1}. The default setting is \code{0.99}. Please adjust for simultaneous testing if pointwise prediction intervals are used for classifying more than one EQA material / reference material in the same IVD-MD comparison.}

\item{rounding}{An \code{integer} specifying the desired decimal places for the predictions and prediction intervals. The default setting is three, offering sufficient precision. The maximum limit is six due to the utilization of floating-point numbers.}
}
\value{
A \code{list} comprising estimated prediction interval data based on user inputs. The output \code{list} may be converted to \code{data.table} by employing the \code{setDT()} method from the \code{data.table} package in R.
}
\description{
This function estimates prediction intervals for External Quality Assessment (EQA) data by applying either Deming (\code{method = 'fg'} or \code{method = 'clsi'}) or Ordinary Least Squares (\code{method = 'ols'}) regression methodologies. It is specifically designed for a single In Vitro Diagnostic Medical Device (IVD-MD) comparison. For processing multiple IVD-MD comparisons simultaneously, consider using the \code{estimate_prediction_data()} function from the commutability package, which executes this function on a per IVD-MD comparison basis.
}
\details{
Prediction Interval Estimation for EQA Data via Deming or OLS Regression

For optimal results, we recommended to include \code{SampleID}, \code{MP_A}, and \code{MP_B} in new_data whenever possible. If only \code{MP_B} is available, the use of \code{method = 'ols'} is not possible. For instance, when constructing prediction bands, only \code{MP_B} might be available. Thus, \code{method = 'ols'} may sadly not be used to estimate pointwise prediction intervals using this function.
}
\examples{
\dontrun{
library(fasteqa)
# Simulation parameters for clinical sample data
training_parameters <- list(n = 25, R = 3, cvx = 0.01, cvy = 0.015, cil = 10, ciu = 70)
# Simulation parameters for external quality assessment material data (commutable materials)
test_parameters <- list(n = 5, R = 3, cvx = 0.01, cvy = 0.015, cil = 10, ciu = 70)
# Simulation of clinical sample data using the simulation parameters 'training_parameters'
training_data <- simulate_eqa_data(training_parameters)
# Simulation of external quality assessment material data using the simulation parameters 'test_parameters'
test_data <- simulate_eqa_data(test_parameters)
# Convert ID columns to character type because this is the type accepted
training_data$SampleID <- as.character(training_data$SampleID)
training_data$ReplicateID <- as.character(training_data$ReplicateID)
# Estimate imprecision estimates
imprecision <- global_precision_estimates(training_data)
# Calculate mean-of-replicates data
mean_of_replicates_training_data <- fun_of_replicates(training_data)
mean_of_replicates_test_data <- fun_of_replicates(test_data)
# Estimate prediction intervals using method = 'fg', R = 3L and R_ratio = 1:
prediction_intervals <- predict_eqa(mean_of_replicates_training_data,
                                    mean_of_replicates_test_data,
                                    imprecision)
}
}
