% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resample_samples}
\alias{resample_samples}
\title{Resample Clustered Data}
\arguments{
\item{data}{A \code{list} or a \code{data.table}. Must contain \code{SampleID},
\code{ReplicateID}, \code{MP_A} and \code{MP_B}. The ID variables \code{SampleID}
and \code{ReplicateID} must be of character type for the function to operate correctly.}
}
\value{
A \code{list} containing the resampled data.
}
\description{
Resample Clustered Data
}
\details{
This function is a very efficient method to resample clinical sample data on sample-level.
It is convenient to combine this function with \code{fasteqa} functions such as
\itemize{
  \item \code{global_precision_estimates()} to estimate bootstrap imprecision confidence intervals
  \item \code{estimate_zeta()} to estimate bootstrap zeta confidence intervals
}
}
\examples{
print(1)
}
