% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{residuals_eqa}
\alias{residuals_eqa}
\title{Calculate Model Residuals for External Quality Assessment Data}
\arguments{
\item{data}{A list or data.table containing:
\itemize{
   \item \code{SampleID}: A \code{character} vector of sample identifiers
   \item \code{MP_A}: A \code{numeric} vector of measurements from method A (response)
   \item \code{MP_B}: A \code{numeric} vector of measurements from method B (predictor)
}}

\item{imprecision_estimates}{A \code{list} or \code{data.table} containing
\itemize{
   \item \code{lambda}: A positive numeric value representing the repeatability
                        variance ratio between response and predictor.
}
Can be calculated using the \code{global_precision_estimates()} function.}

\item{method}{A \code{character} string specifying the linear regression model
applied for obtaining the residuals and fitted values.  Options are:
\itemize{
   \item \code{'fg'} (default): Deming regression, incorporating results from W. Fuller and J. Gillard.
   \item \code{'clsi'}: Deming regression based on the CLSI EP14 standard.
   \item \code{'ols'}: Ordinary Least Squares regression.
}}

\item{studentize}{A non-missing \code{logical} value. If \code{TRUE} (default),
the standardized residuals are computed. Here, this means that the sample mean
of the residuals is subtracted from each residual and this result is divided by
the sample standard deviation of the residuals. Raw residuals are returned
if this is set to \code{FALSE}. Standardized residuals can be
beneficial if one requires to compare the residuals' distribution to the standard
normal distribution or a particular student t-distribution.}

\item{unit_sd}{A non-missing \code{logical} value. If \code{TRUE}, the standard
deviation of the residuals are assumed to be equal to \code{1}.}

\item{invalid_NA}{A non-missing \code{logical} value. If \code{TRUE},
a \code{list} is \code{NA}-values is returned instead of throwing
an error.}
}
\value{
This function returns a \code{list} comprising two \code{numeric} vectors named
\code{residuals} and \code{fitted}. These are the residuals and fitted values, respectively,
based on the input \code{data}, \code{imprecision_estimates}, and \code{method}.
}
\description{
This function computes residuals and fitted values for external quality assessment (EQA)
clinical sample data using various regression methods. It is designed to work with
paired measurement results from two in vitro diagnostic medical devices (IVD-MDs).
}
\details{
The function handles missing values by removing them before calculations. It supports
three regression methods, each with slightly different approaches to calculating residuals
and fitted values. The \code{'fg'} method is generally recommended.
While the outcome differences between the three avaiable methods may be small,
they can still be noteworthy in some cases.
}
\examples{
\dontrun{
  library(fasteqa)
  parameters <- list(n = 20, R = 3, cvx = 0.03, cvy = 0.02, cil = 10, ciu = 15)
  data <- simulate_eqa_data(training_parameters)
  data$SampleID <- as.character(data$SampleID)
  data$ReplicateID <- as.character(data$ReplicateID)
  imprecision <- global_precision_estimates(data)
  mean_of_replicates_data <- fun_of_replicates(data)
  # Extracting raw residuals based on the clsi Deming regression approach
  residuals_eqa(mean_of_replicates_data, imprecision, "clsi", 0)
}

}
\references{
Fuller, W. A. (1987). Measurement Error Models. John Wiley & Sons.

Gillard J. W. (2010). An Overview of Linear Structural Models in Errors in
Variables Regression. REVSTAT-Statistical Journal 8(1), 57-80
}
\seealso{
\code{\link{global_precision_estimates}} for calculating imprecision estimates
}
