% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_eqa_data}
\alias{simulate_eqa_data}
\title{Simulation of EQA data based on study design and potential differences in non-selectivity}
\arguments{
\item{parameters}{A \code{list} of parameters used to simulate the EQA data. You must at least specify one parameter for this function to run. Except that one mandatory parameter, you may optionally choose the remaining of the parameters. These are the optimal parameters that you may include into the list:
\itemize{
  \item{\code{n: }}{  The number of clinical samples}
  \item{\code{R: }}{The number of replicates on each sample}
  \item{\code{cvx: }}{The analytical CV of x measurements}
  \item{\code{cvy: }}{The analytical CV of y measurements}
  \item{\code{cil: }}{The lower range of the concentration interval}
  \item{\code{ciu: }}{The upper range of the concentration interval}
  \item{\code{qpos: }}{Position of systematic differences in non-selectivity. 0 signify lower range and 1 signify upper range}
  \item{\code{qran: }}{Interquantile range where systematic differences in non-selectivity should have its effect}
  \item{\code{prop: }}{average proportion of clinical samples affected by random differences in non-selectivity}
  \item{\code{mmax: }}{The maximum relocation magnitude in number of analytical SDs of y measurements. This assumes either prop or qpos and qran to be specified as well}
}}

\item{silence}{\code{Integer} should temporary calculation results be printed? This may be useful for debugging or strange curiosity. \code{silence = 1} signify that printing will be suppressed, which is the default. \code{silence = 0} allows such printing}
}
\value{
A list where each list element is a column of the generated EQA data
}
\description{
Simulates a data set with n x R rows, and four columns. The two first columns are the base ID columns (\code{SampleID} and \code{ReplicateID}). The remaining columns are numeric columns holding measurement results from the two MSs in comparison (MS_A and MS_B) denoted \code{MP_A} and \code{MP_B}.
}
\details{
Simulation of EQA data based on study design and potential differences in non-selectivity

In order to convert the outputted list to a table, use either \code{as.data.frame()}, \code{as.data.table()}, \code{as.tibble()}. The most efficient way to convert is \code{setDT()} from the \code{data.table} package.
}
\examples{
\dontrun{
  simulated_data <- simulate_eqa_data(parameters = list(n = 25, R = 3, prop = 0.1, mmax = 5))
  simulated_data <- setDT(simulated_data)
}

}
